Use CGFloat for event coordinates until the very end, simplify atomics.
The older atomic logic was probably more correct, but no one should
care if only one coordinate gets updated during a drag, and the new
logic is much simpler.

+++ src/event_tap.hpp
@@ -44,7 +44,7 @@ protected:
         return on_mouse_down(proxy, type, event);
     }
 
-    virtual bool on_left_mouse_drag(CGEventTapProxy proxy, CGEventType type, CGEventRef event, int64_t delta_x, int64_t delta_y) {
+    virtual bool on_left_mouse_drag(CGEventTapProxy proxy, CGEventType type, CGEventRef event, CGFloat delta_x, CGFloat delta_y) {
         return on_mouse_drag(proxy, type, event, delta_x, delta_y);
     }
 
@@ -56,7 +56,7 @@ protected:
         return on_mouse_down(proxy, type, event);
     }
 
-    virtual bool on_right_mouse_drag(CGEventTapProxy proxy, CGEventType type, CGEventRef event, int64_t delta_x, int64_t delta_y) {
+    virtual bool on_right_mouse_drag(CGEventTapProxy proxy, CGEventType type, CGEventRef event, CGFloat delta_x, CGFloat delta_y) {
         return on_mouse_drag(proxy, type, event, delta_x, delta_y);
     }
 
@@ -68,7 +68,7 @@ protected:
         return false;
     }
 
-    virtual bool on_mouse_drag(CGEventTapProxy proxy, CGEventType type, CGEventRef event, int64_t delta_x, int64_t delta_y) {
+    virtual bool on_mouse_drag(CGEventTapProxy proxy, CGEventType type, CGEventRef event, CGFloat delta_x, CGFloat delta_y) {
         return false;
     }
